Attach the recorded-test proposal to the chat input
A plan waiting on a decision belongs with the other things waiting on
one — the questionnaire, the prompt queue, the consent banner — not
scrolled away up the transcript. The proposal now renders on the
composer, and the assistant message keeps a one-line receipt of what the
plan was and what was decided.

Every feature of the old card survives: editing an assertion's sentence,
adding one after a step, removing, drag- and keyboard-reordering,
showing the generated code, approving (which writes the spec and hands
it back to the parked turn), and closing without generating. Spacing is
tighter throughout — the title and filename share the header row — and
because a recording can run long, the timeline is capped at 11rem and
scrolls rather than pushing the message box off the screen. The rows
collapse behind the header; the action row does not, so folding a long
plan away hides the steps, not the decision.

Exactly one of the two renders a given plan at a time: the composer
takes it while the tag still says "proposed", the receipt takes over the
moment it doesn't, so there is never a second approve button or a second
"Generated" badge on screen.

The plan lives in the message rather than in the parked user-input
request, so the composer reads it back out with `listAssertionsTags`,
which scans attributes only and hands back the tag verbatim — the
payload parse is memoized on that string, so restreaming the message
around an open card can't re-seed the rows over the user's edits.
